rename FloatOptions component's name file to same name
